[Chore] #2 UI 수정 및 데이터 타입 변경

+++ BookEntity+CoreDataProperties.swift
@@ -19,7 +19,7 @@ extension BookEntity {
     @NSManaged public var title: String?
     @NSManaged public var author: String?
     @NSManaged public var content: String?
-    @NSManaged public var price: Int32
+    @NSManaged public var price: Int64
     @NSManaged public var thumbnail: String?
 
 }
